feat(sunburst): move stories to the dedicated storybook workspace
