+++ storybook/package.json
@@ -34,6 +34,7 @@
     "@nivo/sankey": "workspace:*",
     "@nivo/scatterplot": "workspace:*",
     "@nivo/stream": "workspace:*",
+    "@nivo/sunburst": "workspace:*",
     "@nivo/tooltip": "workspace:*",
     "@nivo/waffle": "workspace:*",
     "@babel/preset-env": "^7.21.5",
